+++ packages/cubit/lib/src/cubit.dart
@@ -5,7 +5,7 @@ import 'transition.dart';
 
 /// {@template cubit}
 /// A [Cubit] is a subset of [bloc](https://pub.dev/packages/bloc)
-/// which does not contain events and relies on methods to `emit` new states.
+/// which has no notion of events and relies on methods to `emit` new states.
 ///
 /// Every `cubit` requires an `initialState` which will be the
 /// state of the `cubit` before `emit` has been called.