+++ packages/cubit/test/cubits/fake_async_cubit.dart
@@ -3,5 +3,12 @@ import 'package:cubit/cubit.dart';
 class FakeAsyncCounterCubit extends Cubit<int> {
   FakeAsyncCounterCubit() : super(0);
 
-  Future<void> increment() async => emit(state + 1);
+  Future<void> increment() async {
+    final nextState = await _increment(state);
+    emit(nextState);
+  }
+
+  Future<int> _increment(int value) async {
+    return value + 1;
+  }
 }