+++ packages/cubit/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **BREAKING**: update `emit` to be a `void` function
 - **BREAKING**: remove artifical wait to guarantee `initialState` is emitted
+- **BREAKING**: `close` drains the internal `Stream`
 - tests: 100% test coverage
 
 # 0.0.1