+++ packages/cubit/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.0.7
 
 - feat: add `onTransition` to `cubit`
+- feat: `Cubit` extends `CubitStream`
 
 # 0.0.6